Added MealOClient Cell Renderer
Class provides coloring of rows based on the first character found
in column Description. If its a number it colors the row from the colors
array pallete ( supports 10 colors ). If it is any other character than
a number, the class colors the row red ( 11th color in the pallete ).
10 different colors for a meal number from 1 - 10 is enough assuming that 
max of 10 different meals per day will be offered.