Updated MealO class
The database connection is now established before the GUI is shown.
Database connection starts up a network server and tries to get a
embedded connection for this GUI, which makes the data transfer between
it and the database faster. That can sometimes take time and the GUI is
sometimes not shown immediately.