Added a MealO Client Application
This is a second part of the application which provides a user to claim
his order for that day on the shift ( this function however still needs
to be added, more talk about it at the end of the message ). When the 
user inputs his key, the application fetches the order from the database.
If the order was not found, the application adds a row to the table with 
a "Meal Not Ordered" message in each column.
If the order was found and was not yet claimed, the application adds a
row to the table with the corresponding data and adds a timestamp to the
order in the database.
If the order was found and was already claimed, the application add a
row to the table with a "Meal Already Claimed on <date> <time>" message.

More about the order claim on the date and shift:
The application does not yet support a recognition of the correct shift,
which will be added on later time. The application should recognize on
date:
shift 1 when time is 6:01 - 14:00, 
shift 2 when time is 14:01 - 22:00 and
shift 3 when time is 22:01 - 6:00 next day.